test-requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
belfrywidgets==1.0.3
Brotli==1.0.9
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
comiccrawler==2022.2.6
desktop3==0.5.3
discord.py @ file:///data/data/com.termux/files/home/developing/serverbk/ripperv2/test/discord.py
docopt==0.6.2
dropbox==11.15.0
EbookLib==0.17.1
filetype==1.0.10
future==0.18.2
idna==3.2
lxml==4.6.3
lzstring==1.0.4
multidict==5.1.0
node-vm2==0.4.1
Pillow==9.0.1
ply==3.11
pycryptodomex==3.14.1
pyperclip==1.8.2
python-dotenv==0.19.0
python-magic==0.4.25
pythreadworker==0.9.0
requests==2.27.1
safeprint==0.2.0
six==1.16.0
soupsieve==2.2.1
stone==3.2.1
typing-extensions==3.10.0.0
uncurl==0.0.11
urllib3==1.26.6
win-unicode-console==0.5
yarl==1.6.3
